refactor: コードレビュー指摘事項をすべて修正する
- [重大] CheatSheet.tsx: performDrop内の未使用変数 gi を削除
- [重要] CheatSheet.tsx: key={blockIndex/itemIndex} を key={block._editId/item._editId} に修正（DnD再マウント対策）
- [重要] CheatSheet.tsx: 未使用の editFlatIndices useMemo を削除
- [重要] CheatSheet.tsx: RELOAD_CHEAT_SHEET リスナーに editModeRef ガードを追加（編集中の外部リロードを無視）
- [重要] GroupEditDialog.tsx: Enter キーに e.isComposing チェックを追加（IME確定との競合防止）
- [提案] EditIconButton.tsx (atom) を新規作成し、EditableCommandRow/ShortcutRow/GroupBox の IconBtn 重複を統合
- [提案] EditableRowBase.tsx (molecule) を新規作成し、EditableCommandRow/ShortcutRow の共通部分を共通化
- [提案] EditDialog.tsx (molecule) を新規作成し、3ダイアログのslotProps重複を共通化
- [提案] EditBlock.ts: _editCounter（モジュールレベル変数）を crypto.randomUUID() に置換
- CheatSheet.tsx: upsertCommand/upsertGroup の新規ID生成も crypto.randomUUID() に統一
