+++ src/components/atoms/index.ts
@@ -1,4 +1,5 @@
 export * from './AddRowButton'
+export * from './EditIconButton'
 export * from './FileEditButton'
 export * from './FileOpenButton'
 export * from './ThemedSwitch'
