+++ src/components/molecules/CommandEditDialog.tsx
@@ -1,8 +1,9 @@
 'use client'
-import { Box, Dialog, MenuItem, Select, TextField } from '@mui/material'
+import { Box, MenuItem, Select, TextField } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
 import { useEffect, useRef, useState } from 'react'
 
+import { EditDialog } from '@/components/molecules/EditDialog'
 import { FooterButton } from '@/components/molecules/FooterButton'
 import { CommandLayout } from '@/types/api/CheatSheet'
 import { EditCommandData, GroupOption } from '@/types/edit/EditBlock'
@@ -97,33 +98,7 @@ export function CommandEditDialog({
   }
 
   return (
-    <Dialog
-      open
-      onClose={onCancel}
-      slotProps={{
-        backdrop: {
-          sx: {
-            backgroundColor: isDark
-              ? 'rgba(0,0,10,0.45)'
-              : 'rgba(20,30,60,0.28)',
-            backdropFilter: 'blur(3px)',
-          },
-        },
-        paper: {
-          sx: {
-            width: 460,
-            maxWidth: 460,
-            borderRadius: '14px',
-            border: `0.5px solid ${isDark ? 'rgba(255,255,255,0.10)' : 'rgba(0,0,0,0.12)'}`,
-            boxShadow: isDark
-              ? '0 24px 64px rgba(0,0,0,0.65)'
-              : '0 24px 64px rgba(0,0,50,0.30)',
-            overflow: 'hidden',
-            m: 0,
-          },
-        },
-      }}
-    >
+    <EditDialog width={460} onClose={onCancel}>
       {/* ヘッダー */}
       <Box
         sx={{
@@ -197,7 +172,7 @@ export function CommandEditDialog({
               size='small'
               fullWidth
               onKeyDown={(e) => {
-                if (e.key === 'Enter' && canSave) handleSave()
+                if (e.key === 'Enter' && canSave && !e.nativeEvent.isComposing) handleSave()
               }}
             />
           </FieldRow>
@@ -255,7 +230,7 @@ export function CommandEditDialog({
           Save
         </FooterButton>
       </Box>
-    </Dialog>
+    </EditDialog>
   )
 }
