+++ src/components/molecules/index.ts
@@ -3,7 +3,9 @@ export * from './CommandField'
 export * from './DeleteConfirmDialog'
 export * from './EditableCommandRow'
 export * from './EditableGroupBox'
+export * from './EditableRowBase'
 export * from './EditableShortcutRow'
+export * from './EditDialog'
 export * from './EditModeFooter'
 export * from './FooterButton'
 export * from './GroupEditDialog'
